restore: require --source when combining --worktree and --staged
The default restore source for --worktree is the index, and the default
source for --staged is HEAD. When combining --worktree and --staged in
the same invocation, the restore source is ambiguous ("should it restore
from the index or from HEAD?"). To avoid such ambiguity, the git-restore
documentation has always stated that --source must be used when
combining --worktree and --staged. However, this restriction is not
actually enforced. Address this deficiency by making the implementation
match the documented behavior (to wit, error out if --source is not
specified when combining --worktree and --staged).

While at it, enhance the documentation to mention the --source
requirement prominently in the "Description" section (rather than only
in the description of the --source option itself).

Reported-by: Harri Mehtälä <harri.mehtala@finago.com>
Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>