restore: default to HEAD when combining --worktree and --staged
The default restore source for --worktree is the index, and the default
source for --staged is HEAD. However, when combining --worktree and
--staged in the same invocation, git-restore requires the source to be
specified explicitly via --source since it would otherwise be ambiguous
("should it restore from the index or from HEAD?"). This requirement
makes it cumbersome to restore a file in both the worktree and the
index.

However, HEAD is also a reasonably intuitive default restore source when
--worktree and --staged are combined. After all, if a user is asking to
throw away all local changes to a file (on disk and in the index)
without specifying a restore source explicitly -- and the user expects
the file to be restored from _somewhere_ -- then it is likely that the
user expects them to be restored from HEAD, which is an intuitive and
logical place to find a recent unadulterated copy of the file.

Therefore, make HEAD the default restore source when --worktree and
--staged are combined.

Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>